change sidebar ui on desktop
Was using the Drawer component for the sidebar for both small and large screens, but didnt work that well on desktop
and covered parts of map since it has to be positioned absolutely. Switched to just a div that can have
its width toggled to 0 with a button. Map width grows to fill remaining space